backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==0.21.0
google-generativeai==0.8.3
mcp
aiohttp==3.11.11
requests==2.32.3
pydantic==2.10.1

